Refactor encoding scheme
This commit introduces a significant refactor to how the expert encoding
scheme works, and other subsequent changes around this. The motivation
for this change is as follows:

By encoding the samples as we use them, instead of sampling from one
monolithic enocding produced at the beginning of training, we are
now actually leveraging the stochasticity of the world model, and thus,
fixing the overfitting issue which has been present for months.

The specific changes can be broken down as follows:
1. The ditto step now takes the expert sampler directly
   as an arguement and encodes the data on the fly
2. encode_expert_data() has been made redundant and removed. All encoding
   functionality is now moved instead encode_data()
3. The only storage class needed now is the ExpertSampler, into which a
   sample method has been added
4. The inputs to function in ditto.py and visualizer.py have been
   subsequently updated
5. train_step() has been commented out. N.B. dreamer functionality is
   broken and needs fixing
6. PEP8 fixes to distributions.py