Refactor file structure
The file structure of the repo has been changed to be more in line with
best practicies.
The major changes can be broken down as follows:
1. world_models module has been renamed to dreamer, and now only
contrains files that are a part of the module.
2. raisim_gym has been refactored into the usual include/src format
3. data and logs have been moved to the root directory
4. ditto training is now done by running ditto.py

N.B. train.py and train_dreamer.py are broken and need to be fixed or
deleted, depending on future plans regarding optional dreamer mode